+++ src/tuple/t_tuple_multi_scal.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   t_tuple_multi_scal.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: tda-silv <tda-silv@student.42.fr>          +#+  +:+       +#+        */
+/*   By: roberto <roberto@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/18 17:12:44 by tda-silv          #+#    #+#             */
-/*   Updated: 2023/02/21 23:43:25 by tda-silv         ###   ########.fr       */
+/*   Updated: 2023/03/08 21:28:35 by roberto          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 /*   Multiplie un tuple par un scalaire.									  */
 /*                                                                            */
 /* ************************************************************************** */
-t_tuple	t_tuple_multi_scal(t_tuple left, float right)
+t_tuple	t_tuple_multi_scal(t_tuple left, double right)
 {
 	t_tuple	ret;
